Change to make this work on 64 bit machines.
git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk@5207 640d5228-2204-0410-8d71-8f46fa6850b4

+++ nc_utils/ncReorder/ncReorder.cc
@@ -120,7 +120,7 @@ int main(int argc, char *argv[])
 
         case ncInt:
           outFile.get_var(i)->add_att(att->name(), att->num_vals(),
-		(const long *)att->values()->base());
+		(const int *)att->values()->base());
           break;
 
         default:
